% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{G_priorDensity}
\alias{G_priorDensity}
\title{Plot Dirichlet / Pitman-Yor process Priors}
\usage{
G_priorDensity(N, alpha, discount = 0L, show.plot = TRUE, col = NULL,
  main = NULL, ...)
}
\arguments{
\item{N}{The sample size.}

\item{alpha}{The concentration parameter. Must be specified and must be strictly greater than \code{-discount}.}

\item{discount}{The discount parameter for the Pitman-Yor process. Must lie in the interval [0, 1). Defaults to 0 (i.e. the Dirichlet process).}

\item{show.plot}{Logical indicating whether the plot should be displayed (default = \code{TRUE}).}

\item{col}{Colour of the plotted lines.}

\item{main}{Title of the plot.}

\item{...}{Other optional arguments typically passed to \code{\link{plot}}.}
}
\value{
A plot of the prior distribution if \code{show.plot} is \code{TRUE}. Density values are returned invisibly. Note that the density values may not strictly sum to one in certain cases, as values small enough to be represented as zero may well be returned.
}
\description{
Plots the prior distribution of the number of clusters under a Dirichlet / Pitman-Yor process prior, for a sample of size \code{N} at given values of the concentration parameter \code{alpha} and optionally also the \code{discount} parameter. Useful for soliciting sensible priors for \code{alpha} or suitable fixed values for \code{alpha} or \code{discount} under the "\code{IMFA}" and "\code{IMIFA}" methods for \code{\link{mcmc_IMIFA}}. All arguments are vectorised. Requires use of the \code{Rmpfr} and \code{gmp} libraries. Users can also consult \code{\link{G_expected}} and \code{\link{G_variance}} in order to solicit sensible priors.
}
\examples{
# install.packages("Rmpfr")

# Verify that these alpha/discount values produce priors with the same mean
# G_expected(N=50, alpha=c(19.23356, 12.21619, 1), discount=c(0, 0.25, 0.7300045))

# Now plot them to examine tail behaviour as discount increases
# x <- G_priorDensity(N=50, alpha=c(19.23356, 12.21619, 1), discount=c(0, 0.25, 0.7300045), col=2:4)
# x
}
\seealso{
\code{\link{G_expected}}, \code{\link{G_variance}}, \code{\link[Rmpfr]{Rmpfr}}
}
