% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullConditionals.R
\name{IMIFA_news}
\alias{IMIFA_news}
\title{Show the NEWS file}
\usage{
IMIFA_news(what = c("html", "md"))
}
\arguments{
\item{what}{A character string specifying whether the \code{"html"} (default) or Markdown \code{"md"} format is to be shown.}
}
\value{
The \code{IMIFA} \code{NEWS} file in the desired format.
}
\description{
Show the \code{NEWS} file of the \code{IMIFA} package.
}
\examples{
IMIFA_news()
}
\keyword{utility}
