% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{is.cols}
\alias{is.cols}
\title{Check for Valid Colours}
\usage{
is.cols(cols)
}
\arguments{
\item{cols}{A vector of colours, usually as a character string.}
}
\value{
A logical vector of length \code{length(cols)} which is \code{TRUE} for entries which are valid colours and \code{FALSE} otherwise.
}
\description{
Checks if the supplied vector contains valid colours.
}
\examples{
all(is.cols(1:5))

all(is.cols(heat.colors(30)))

any(!is.cols(c("red", "green", "aquamarine")))
}
\keyword{utility}
