% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullConditionals.R
\name{mixFac_free}
\alias{mixFac_free}
\title{Estimate the Number of Free Parameters in a Finite Factor Analysis (Mixture) Model}
\usage{
mixFac_free(Q, P, G = 1, uni = c("unconstrained", "isotropic"))
}
\arguments{
\item{Q}{The number of latent factors (which can be 0, corresponding to a model with diagonal covariance). This argument is vectorised.}

\item{P}{The number of variables.}

\item{G}{The number of groups. This defaults to 1.}

\item{uni}{By default, calculation assumes the \emph{UUU} model, with "\code{unconstrained}" uniquesses. It's also possible to calculate this quantity for the \emph{UUC} model, with "\code{isotropic}" uniquenesses.}
}
\value{
A vector of length \code{length(Q)}.
}
\description{
Estimates the dimension of the 'free' parameters in \emph{UUU} or \emph{UUC} factor analysis models, used to calculate the penalty terms for the \code{aic.mcmc} and \code{bic.mcmc} model selection criteria implemented in \code{\link{get_IMIFA_results}}. Please note that while this available as a standalone function, no checks are performed in order to make its use in \code{\link{get_IMIFA_results}} faster.
}
\examples{
mixFac_free(Q=4, P=50, G=3, uni="unconstrained")

mixFac_free(Q=3:6, P=100, uni="isotropic")
}
\references{
McNicholas, P. D. and Murphy, T. B. (2008) Parsimonious Gaussian Mixture Models, \emph{Statistics and Computing}, 18(3): 285-296.
}
