% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostics.R
\name{get_IMIFA_results}
\alias{get_IMIFA_results}
\title{Extract results, conduct posterior inference and compute performance metrics for MCMC samples of models from the IMIFA family}
\usage{
get_IMIFA_results(sims = NULL, burnin = 0L, thinning = 1L, G = NULL,
  Q = NULL, criterion = c("bicm", "aicm", "bic.mcmc", "aic.mcmc",
  "log.iLLH", "dic"), G.meth = c("mode", "median"), Q.meth = c("mode",
  "median"), dat = NULL, conf.level = 0.95, zlabels = NULL)
}
\arguments{
\item{sims}{An object of class "\code{IMIFA}" generated by \code{\link{mcmc_IMIFA}}.}

\item{burnin}{Optional additional number of iterations to discard. Defaults to 0.}

\item{thinning}{Optional interval for extra thinning to be applied. Defaults to 1.}

\item{G}{If this argument is not specified, results will be returned with the optimal number of clusters. If different numbers of clusters were explored in \code{sims} for the "\code{MFA}" or "\code{MIFA}" methods, supplying an integer value allows pulling out a specific solution with \code{G} clusters, even if the solution is sub-optimal. Similarly, this allows retrieval of samples corresponding to a solution, if visited, with \code{G} clusters for the "\code{OMFA}", "\code{OMIFA}", "\code{IMFA}" and "\code{IMIFA}" methods.}

\item{Q}{If this argument is non specified, results will be returned with the optimal number of factors. If different numbers of factors were explored in \code{sims} for the "\code{FA}", "\code{MFA}", "\code{OMFA}" or "\code{IMFA}" methods, this allows pulling out a specific solution with \code{Q} factors, even if the solution is sub-optimal. Similarly, this allows retrieval of samples corresponding to a solution, if visited, with \code{Q} factors for the "\code{IFA}", "\code{MIFA}", "\code{OMIFA}" and "\code{IMIFA}" methods.}

\item{criterion}{The criterion to use for model selection, if model selection is required by the method in \code{sims}. Note that these are \emph{all} calculated, this argument merely indicates which one will form the basis of the construction of the output.}

\item{G.meth}{If the object in \code{sims} arises from the "\code{OMFA}", "\code{OMIFA}", "\code{IMFA}" or "\code{IMIFA}" methods, this argument determines whether the optimal number of clusters is given by the mode or median of the posterior distribution of \code{G}. Defaults to "\code{Mode}".}

\item{Q.meth}{If the object in \code{sims} arises from the "\code{IFA}", "\code{MIFA}", "\code{OMIFA}" or "\code{IMIFA}" methods, this argument determines whether the optimal number of latent factors is given by the mode or median of the posterior distribution of \code{Q}. Defaults to "\code{Mode}".}

\item{dat}{The actual data set on which \code{\link{mcmc_IMIFA}} was originally run. This is necessary for computing error metrics between the estimated and empirical covariance matrix/matrices. If this is not supplied, the function will attempt to find the data set if it is still available in the global environment.}

\item{conf.level}{The confidence level to be used throughout for credible intervals for all parameters of inferential interest. Defaults to 0.95.}

\item{zlabels}{For any method that performs clustering, the true labels can be supplied if they are known in order to compute clustering performance metrics. This also has the effect of ordering the MAP labels (and thus the ordering of cluster-specific parameters) to most closely correspond to the true labels if supplied.}
}
\value{
An object of class "\code{Results_IMIFA}" to be passed to \code{\link{plot.Results_IMIFA}} for visualising results. Dedicated \code{print} and \code{summary} functions exist for objects of this class. The object is a list of lists, some of the most important components of which are:
\describe{
\item{Clust}{Everything pertaining to clustering performance can be found here for all but the "\code{FA}" and "\code{IFA}" methods. More detail is given if \code{zlabels}} are supplied.
\item{Error}{Error metrics (e.g. MSE) between the empirical and estimated covariance matrix/matrices.}
\item{GQ.results}{Everything pertaining to model choice can be found here, incl. posterior summaries for the estimated number of groups and estimated number of factors, if applicable to the method employed. Information criterion values are also accessible here.}
\item{Means}{Posterior summaries for the means.}
\item{Loadings}{Posterior summaries for the factor loadings matrix/matrices.}
\item{Scores}{Posterior summaries for the latent factor scores.}
\item{Uniqueness}{Posterior summaries for the uniquenesses.}
}
}
\description{
This function post-processes simulations generated by \code{\link{mcmc_IMIFA}} for any of the IMIFA family of models. It can be re-ran at little computational cost in order to extract different models explored by the sampler used for \code{sims}, without having to re-run the model itself. New results objects using different numbers of clusters and different numbers of factors (if visited by the model in question), or using different model selection criteria (if necessary) can be generated with ease. The function also performs post-hoc corrections for label switching, as well as post-hoc Procrustes rotation, to ensure sensible posterior means
}
\examples{
# data(coffee)
# data(olive)

# Run a MFA model on the coffee data over a range of clusters and factors.
# simMFAcoffee  <- mcmc_IMIFA(coffee, method="MFA", range.G=2:3, range.Q=0:3, n.iters=1000)

# Accept all defaults to extract the optimal model.
# resMFAcoffee  <- get_IMIFA_results(simMFAcoffee)


# Instead let's produce results for a 3-cluster model, allowing the number of factors to be chosen by \\code{aic.mcmc}.
# resMFAcoffee2 <- get_IMIFA_results(simMFAcoffee, G=3, criterion="aic.mcmc")

# Run an IMIFA model on the olive data, accepting all defaults.
# simIMIFAolive <- mcmc_IMIFA(olive, method="IMIFA", n.iters=10000)

# Extract the optimum results, estimated \\code{G} and \\code{Q} by the \\emph{median} of their posterior distributions, and construct 90\% credible intervals.
# resIMIFAolive <- get_IMIFA_results(simIMIFAolive, G.meth="median", Q.meth="median", conf.level=0.9)
}
\references{
Murphy, K., Gormley, I.C. and Viroli, C. (2017) Infinite Mixtures of Infinite Factor Analysers: Nonparametric Model-Based Clustering via Latent Gaussian Models, \code{https://arxiv.org/abs/1701.07010}
}
\seealso{
\code{\link{mcmc_IMIFA}}, \code{\link{plot.Results_IMIFA}}
}

