% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullConditionals.R
\name{IMIFA_news}
\alias{IMIFA_news}
\title{Show the NEWS file}
\usage{
IMIFA_news()
}
\value{
The \code{IMIFA} \code{NEWS} file, provided the session is interactive.
}
\description{
Show the \code{NEWS} file of the \code{IMIFA} package.
}
\examples{
IMIFA_news()
}
\keyword{utility}
