% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{mat2cols}
\alias{mat2cols}
\title{Convert a numeric matrix to colours}
\usage{
mat2cols(mat, cols = heat.colors(30L), byrank = FALSE,
  breaks = length(cols), blind = TRUE, blind.type = c("deutan", "protan",
  "tritan"))
}
\arguments{
\item{mat}{A matrix.}

\item{cols}{The colour palette to be used. Defaults to \code{heat.colors(30)}. Will be checked for validity.}

\item{byrank}{Logical indicating whether to convert the matrix itself or the sample ranks of the values therein. Defaults to FALSE.}

\item{breaks}{Number of gradations in colour to use. Defaults to \code{length(cols)}.}

\item{blind}{A logical switch indicating whether colour blind plots using \code{\link[dichromat]{dichromat}} should be produced. Defaults to TRUE.}

\item{blind.type}{The type of colour palette to be used if \code{blind} is TRUE. The arguments are exactly as the \code{type} argument under \code{\link[dichromat]{dichromat}}.}
}
\value{
A matrix of hex colour code representations.
}
\description{
Converts a matrix to a hex colour code representation for plotting using \code{\link[gclus]{plotcolors}}. Used internally by \code{\link{plot.Results_IMIFA}} for plotting posterior mean loadings heatmaps.
}
\examples{
mat  <- matrix(rnorm(100), nr=10, nc=10)
cols <- mat2cols(mat)
cols
plotcolors(cols)
# plotcolors(cols, blind=FALSE, breaks=10)
}
\seealso{
\code{\link[gclus]{plotcolors}}
}

